BUGFIX-5: Extract Windows platform size polling function into a separate module
